+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
 
+## 0.4.2
+
+Released on January 30, 2024.
+
+### Added
+* Tensor input support for FaceAnalyzer run method
+* TensorReader for transforming incoming torch tensors
+
+
 ## 0.4.1
 
 Released on December 14, 2023.